requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10
pydantic==2.11.9
PyJWT==2.10.1
cryptography==45.0.7
python-dotenv==1.1.1
httpx==0.28.1
haversine==2.9.0
openai>=1.0.0
